backend/pyproject.toml
[tool.poetry]
name = "diet"
version = "0.1.0"
description = ""
authors = ["masfana016 <masfanasrullahansari123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
langchain-core = "^0.3.21"
langchain-community = "^0.3.9"
langchain-google-genai = "^2.0.6"
langgraph = "^0.2.53"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.9.0.post1"
beautifulsoup4 = "^4.12.3"
sqlmodel = "^0.0.22"
bcrypt = "^4.2.1"
psycopg2-binary = "^2.9.10"
pydentic = "^0.0.1.dev3"
pydantic = {extras = ["email"], version = "^2.10.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
